Threading: Use 'size_t' for stack size in core threading code
For some reason it was defined as 'unsigned short' but all arguments to
the threading functions and other structs used size_t.  The SDL plugin
tried to allocate a 2MB stack and this resulted in much badness.

This is a _very_ old bug, and might be responsible for countless subtle
bugs!

+++ firmware/kernel/thread-internal.h
@@ -128,7 +128,7 @@ struct thread_entry
                                     all threads blocked by this one) */
 #endif
 #ifndef HAVE_SDL_THREADS
-    unsigned short stack_size;   /* Size of stack in bytes */
+    size_t stack_size;           /* Size of stack in bytes */
 #endif
     unsigned char state;         /* Thread slot state (STATE_*) */
 #ifdef HAVE_SCHEDULER_BOOSTCTRL
